Reimplement as part of KeyEventChannel
