+++ shell/platform/tizen/BUILD.gn
@@ -126,7 +126,6 @@ template("embedder") {
       "flutter_tizen_engine.cc",
       "flutter_tizen_texture_registrar.cc",
       "flutter_tizen_view.cc",
-      "key_event_handler.cc",
       "key_mapping.cc",
       "logger.cc",
       "system_utils.cc",
