+++ shell/platform/tizen/flutter_tizen_engine.cc
@@ -238,7 +238,10 @@ bool FlutterTizenEngine::RunEngine() {
   if (IsHeaded()) {
     texture_registrar_ = std::make_unique<FlutterTizenTextureRegistrar>(this);
     key_event_channel_ = std::make_unique<KeyEventChannel>(
-        internal_plugin_registrar_->messenger());
+        internal_plugin_registrar_->messenger(),
+        [this](const auto& event, auto callback, void* user_data) {
+          SendKeyEvent(event, callback, user_data);
+        });
     navigation_channel_ = std::make_unique<NavigationChannel>(
         internal_plugin_registrar_->messenger());
     platform_view_channel_ = std::make_unique<PlatformViewChannel>(
